+++ R/select_col.R
@@ -8,7 +8,7 @@
 #' @return An object of the same type as the input data.
 #' @references This function makes use of the \code{\link[dplyr]{select}} function
 #'             within the \code{dplyr} package.
-#' @seealso \code{\link{rtry_rm_col}}
+#' @seealso \code{\link{rtry_remove_col}}
 #' @examples
 #' # Select certain columns from the provided sample data (TRYdata_15160)
 #' data_selected <- rtry_select_col(TRYdata_15160,